Saved encounters now connected to the current user.

+++ backend/api/routes/encounters.py
@@ -1,10 +1,11 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, HTTPException, status, Depends
 from pydantic import BaseModel
 from sqlalchemy.future import select
 
 import models
 
 from ..dependencies import db_dependency
+from ..auth_helpers import get_current_active_user
 
 router = APIRouter()
 
@@ -20,8 +21,9 @@ class Encounters(BaseModel):
 
 
 @router.get("/encounters", response_model=Encounters, status_code=status.HTTP_200_OK)
-def get_encounters(db: db_dependency):
+def get_encounters(db: db_dependency, current_user: models.User = Depends(get_current_active_user)):
     query = select(models.Encounter)
+    query = query.where(models.Encounter.user_id == current_user.id)
     result = db.execute(query)
     encounters = result.scalars().all()
     encounter_list = [e.__dict__ for e in encounters]
@@ -42,11 +44,12 @@ def get_encounter(encounter_id, db: db_dependency):
 
 
 @router.post("/encounters", response_model=Encounter, status_code=status.HTTP_201_CREATED)
-def add_encounter(encounter: Encounter, db: db_dependency):
+def add_encounter(encounter: Encounter, db: db_dependency, current_user: models.User = Depends(get_current_active_user)):
     try:
         db_encounter = models.Encounter(
             name=encounter.name,
-            enemies=encounter.enemies
+            enemies=encounter.enemies,
+            user=current_user
         )
         db.add(db_encounter)
         db.commit()
@@ -62,14 +65,20 @@ def add_encounter(encounter: Encounter, db: db_dependency):
 
 
 @router.delete("/encounters/{encounter_id}", response_model=object, status_code=status.HTTP_200_OK)
-def delete_encounter(encounter_id, db: db_dependency):
+def delete_encounter(encounter_id, db: db_dependency, current_user: models.User = Depends(get_current_active_user)):
     query = db.query(models.Encounter).where(models.Encounter.id == encounter_id)
 
     result = db.execute(query)
     encounter = result.scalars().first()
 
     if not encounter:
-        raise HTTPException(status_code=404, detail="Encounter not found")
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Encounter not found")
+
+    if encounter.user_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Not authorized to delete this encounter",
+        )
 
     db.delete(encounter)
     db.commit()
